# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend monolith modular untuk platform jasa informal Kerjantara.id. API ini menghubungkan employer dengan worker melalui matching engine berbasis GPS dan skor reputasi.",
        "title": "Kerjantara Backend API",
        "contact": {},
        "version": "v3.3-hackathon"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/ktp/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan semua pengajuan verifikasi KTP dengan status 'pending'. Endpoint ini memerlukan role admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Ambil daftar verifikasi KTP yang menunggu review",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/admin/ktp/{user_id}/review": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin menyetujui, menolak, atau meminta pengiriman ulang dokumen KTP pengguna. Endpoint ini memerlukan role admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Review pengajuan verifikasi KTP pengguna",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID pengguna (UUID)",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Keputusan review",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.ReviewVerificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/google": {
            "post": {
                "description": "Menerima access token dari Supabase (hasil login Google), memverifikasi token, lalu mengembalikan JWT app. User harus sudah terdaftar di mst_users dengan ID yang sama dengan sub claim token Supabase.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login menggunakan Google via Supabase",
                "parameters": [
                    {
                        "description": "Access token dari Supabase",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.GoogleLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_auth.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/ktp/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menerima dua file gambar: foto KTP dan foto selfie. Batas ukuran masing-masing 10MB. Setelah upload, status verifikasi menjadi 'pending' hingga direview admin.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Upload foto KTP dan selfie untuk verifikasi identitas",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Foto KTP (max 10MB)",
                        "name": "ktp_photo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Foto selfie dengan KTP (max 10MB)",
                        "name": "selfie_photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_auth.KTPUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Autentikasi pengguna menggunakan nomor telepon dan password. Mengembalikan JWT Bearer Token yang berlaku 7 hari.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login pengguna",
                "parameters": [
                    {
                        "description": "Kredensial login",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_auth.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan data profil lengkap pengguna yang sedang login berdasarkan JWT token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Ambil profil pengguna saat ini",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Membuat akun pengguna baru dengan role worker atau employer. Token JWT langsung dikembalikan setelah registrasi berhasil.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi pengguna baru",
                "parameters": [
                    {
                        "description": "Data registrasi",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_auth.RegisterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/roles/activate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan role baru (worker atau employer) ke akun pengguna. Mengembalikan token JWT baru yang mencerminkan role aktif.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Aktifkan role baru untuk pengguna",
                "parameters": [
                    {
                        "description": "Role yang ingin diaktifkan",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.ActivateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/docs.ErrorDetail"
                }
            }
        },
...
        "docs.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/docs.MetaInfo"
                }
            }
        },
        "internal_auth.ActivateRoleRequest": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string",
                    "example": "worker"
                }
            }
        },
        "internal_auth.GoogleLoginRequest": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
        "internal_auth.ReviewVerificationRequest": {
            "type": "object",
            "properties": {
                "decision": {
                    "type": "string",
                    "enum": [
                        "approved",
                        "rejected",
                        "resubmit"
                    ],
                    "example": "approved"
                },
                "note": {
                    "type": "string",
                    "example": "KTP terlihat jelas dan valid"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dalam format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}